requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
requests
python-dotenv
fpdf2
jinja2
redis
celery
razorpay
passlib[bcrypt]
python-jose[cryptography]
structlog
pydantic-settings
python-multipart
email-validator
openpyxl
